Port course search from CARPI Bot
We now have a working endpoint that takes in a single search term parameter, and returns a list of raw CourseSearchResult entities.

Not the final product by the way; CourseSearchResult contains some extraneous fields (pkCourses, codeMatch, titleExactMatch, etc.) that the client doesn't need.

I want to make some mapper class that'll be able to map the relevant fields from a CourseSearchResult to a Course entity or a proper DTO. Afterwards, I don't think anything really needs to be done to the search algorithm.