Add department & semester filter to course search
The course DTO now includes information about the semesters it's been offered in, and a list of it's special attributes (HASS Inquiry, etc.).

Issues that need to be worked on:

Attributes haven't been added to the database at this point in time, so the attribute filter has been commented out.

Only a single department filter, attribute filter, and semester filter can be applied. There is currently no support for multiple department filters, etc.

Co-Authored-By: Jack Zgombic <69125339+jzgom067@users.noreply.github.com>