Add endpoints to expose filter values
The filter feature on the website needs to know what filters are available to be used, and this information should be served from the backend in case the data changes on the server side.

Because of Spring Boot things, this commit looks a LOT larger than it should be for such a small addition. In short, it's not so simple to query for just one column of a table and serve it to the frontend; there should be proper entities, repositories, composite keys, and a new thing called projections associated with each table involved.

Projections are basically views of entities; i.e. I wanted to select all distinct departments from the database, but I didn't want to get that data in the form of Course entities. So I created a CourseDepartmentProjection that's compatible with the Course repository and only captures the data that I need. Although it feels kind of stupid making a projection to serve a single String.