Add global exception handlers for controllers
With this change, the ugly default "Whitelabel Error Page" should no longer show, as the new controller advice method handleException() will catch any error and respond with a friendly JSON instead.

There is one other exception handler, handleNoHandlerFoundException() which fires mostly in the case where an unmapped endpoint is accessed. This one just returns a 404 with a "Resource not found" message instead.

I think it's good to have this class in case we want to handle any specific exceptions in the future.

Co-Authored-By: anthonyrys <99291131+anthonyrys@users.noreply.github.com>